Break keys/configuration into separate handlers
The "everything is magically mounted under the discovery endpoint" is somewhat
confusing and inflexible, and the mux handling of slashes/prefixes gets
brittle.

Just split them into two separate handlers for simplicity.
